write a test for each file
