+++ package.json
@@ -7,7 +7,7 @@
     "start": "webpack-dev-server",
     "webpack": "./node_modules/webpack/bin/webpack.js",
     "build": "npm run lint; npm test; npm run webpack",
-    "test": "mocha --compilers js:babel-core/register",
+    "test": "mocha --compilers js:babel-core/register --require test/helpers/setup.js --recursive",
     "lint": "./node_modules/eslint/bin/eslint.js  --ext .js --ext .jsx lib/** __tests__**"
   },
   "repository": {
@@ -46,5 +46,8 @@
     "style-loader": "^0.13.1",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
+  },
+  "dependencies": {
+    "react-addons-test-utils": "^15.4.1"
   }
 }
